backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
pydantic-settings
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
email-validator==2.2.0
bcrypt==4.0.1
python-multipart==0.0.9
torch
python-multipart
argon2-cffi
geoip2
httpx
numpy
scikit-learn
matplotlib